Compiles and runs
