+++ src/esp32/stub/src/cpu_start.c
@@ -3,9 +3,11 @@
 #include <stdbool.h>
 
 extern void app_main(void);
+extern void vTaskStartScheduler();
 jmp_buf buf;
 
 int main(void) {
     setjmp(buf);
     app_main();
+    vTaskStartScheduler();
 }
