+++ src/freertos/stub/CMakeLists.txt
@@ -1,51 +1,52 @@
 set(LIB_NAME "stub_freertos")
+set(LIB_ALIAS "SwarmUs::freertos")
+
 
 include(freertos/common)
 freertos_fetch_kernel()
 
-set(THREADS_PREFER_PTHREAD_FLAG ON)
-find_package(Threads REQUIRED)
-find_package(FreeRTOS COMPONENTS POSIX REQUIRED)
-
-target_compile_options(FreeRTOS::POSIX INTERFACE -fPIC)
-
-
-set(LIB_HEADERS
-    include/freertos/
-)
+add_subdirectory(config)
 
 set(LIB_SOURCES
-    src/platform_freertos_hooks.c
+    src/freertos_hooks.c
 )
 
-add_library(${LIB_NAME} STATIC
-        ${LIB_SOURCES}
+set(LIB_HEADERS
+    include/freertos/FreeRTOSConfig.h
 )
 
-target_link_libraries(${LIB_NAME}
-    PUBLIC
-        Threads::Threads
-        FreeRTOS::POSIX
-        FreeRTOS::Coroutine
-        FreeRTOS::EventGroups
-        FreeRTOS::StreamBuffer
-        FreeRTOS::Timers
-        FreeRTOS::Heap::3
-)
+add_library(${LIB_NAME} STATIC ${LIB_SOURCES} ${LIB_HEADERS})
+add_library(${LIB_ALIAS} ALIAS ${LIB_NAME})
 
+find_package(FreeRTOS COMPONENTS
+        NONE
+        REQUIRED)
 
 target_include_directories(${LIB_NAME}
         PUBLIC
-        ${LIB_HEADERS}
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
 
         # Allows freertos to build properly
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/freertos>
         $<INSTALL_INTERFACE:include>
         ${FREERTOS_KERNEL_PATH}/include
-        ${FREERTOS_KERNEL_PATH}/portable/ThirdParty/GCC/Posix)
+        )
 
+target_link_libraries(${LIB_NAME}
+        PUBLIC
+        SwarmUs::freertos::posix::config
 
-add_library(SwarmUs::freertos ALIAS ${LIB_NAME})
+        PRIVATE
+        SwarmUs::freertos::posix
+
+        FreeRTOS::Coroutine
+        FreeRTOS::EventGroups
+        FreeRTOS::StreamBuffer
+        FreeRTOS::Timers
+        FreeRTOS::Heap::3
+
+        )
+
 
 # Removing warnings from freertos compilation on executable target
 if (DISABLE_EXTERNAL_WARNINGS)
