+++ src/freertos/stub/config/include/freertos_platform_config.h
@@ -1,0 +1,8 @@
+#ifndef __FREERTOS_PLATFORM_CONFIG_H_
+#define __FREERTOS_PLATFORM_CONFIG_H_
+
+/* Clock config that is platform dependent */
+
+#define configTICK_RATE_HZ ((TickType_t)1000)
+
+#endif // __FREERTOS_PLATFORM_CONFIG_H_
